mudando forma tratar balance de float 64 para centavos int64 previnir bugs

+++ entity/structs.go
@@ -4,7 +4,7 @@ import "time"
 
 type ParamCreate struct {
 	AccountID string
-	Balance   float64
+	Balance   int64
 	CreatedAT time.Time
 	UpdatedAT time.Time
 }
@@ -12,21 +12,21 @@ type ParamCreate struct {
 type ParamCreateOutput struct {
 	WalletID  string
 	AccountID string
-	Balance   float64
+	Balance   int64
 	CreatedAT time.Time
 	UpdatedAT time.Time
 }
 
 type ParamUpdate struct {
 	WalletID  string
-	Balance   float64
+	Balance   int64
 	UpdatedAT time.Time
 }
 
 type ParamUpdateOutput struct {
 	WalletID  string
 	AccountID string
-	Balance   float64
+	Balance   int64
 	CreatedAT time.Time
 	UpdatedAT time.Time
 }
@@ -38,7 +38,7 @@ type ParamGet struct {
 type ParamGetOutput struct {
 	WalletID  string
 	AccountID string
-	Balance   float64
+	Balance   int64
 	CreatedAT time.Time
 	UpdatedAT time.Time
 }
@@ -50,7 +50,7 @@ type ParamGetByAccount struct {
 type ParamGetByAccountOutput struct {
 	WalletID  string
 	AccountID string
-	Balance   float64
+	Balance   int64
 	CreatedAT time.Time
 	UpdatedAT time.Time
 }
