+++ infra/repository/dto.go
@@ -9,7 +9,7 @@ import (
 type paramRepositoryMongoInput struct {
 	WalletID  primitive.ObjectID `bson:"_id,omitempty"`
 	AccountID string             `bson:"account_id"`
-	Balance   float64            `bson:"balance"`
+	Balance   int64            `bson:"balance"`
 	CreatedAT time.Time          `bson:"created_at"`
 	UpdatedAT time.Time          `bson:"updated_at"`
 }
@@ -17,7 +17,7 @@ type paramRepositoryMongoInput struct {
 type paramRepositoryMongoOutput struct {
 	WalletID  primitive.ObjectID `bson:"_id,omitempty"`
 	AccountID string             `bson:"account_id"`
-	Balance   float64            `bson:"balance"`
+	Balance   int64            `bson:"balance"`
 	CreatedAT time.Time          `bson:"created_at"`
 	UpdatedAT time.Time          `bson:"updated_at"`
 }
